Reduce IDLE timeout.
Summary:
This seems like the most expedient way to ensure that we detect flag changes
with at least reasonable latency.

Fixes T319.

Test Plan:
Mark a thread as read in the Gmail web UI; check that this gets
propagated to the API within a reasonable amount of time if (and only if) a
short idle timeout is set.

Reviewers: spang

Reviewed By: spang

Maniphest Tasks: T319

Differential Revision: https://review.inboxapp.com/D387