[api/drafts] Accept PUT to /drafts/ID instead of POST.
Summary:
According to Wikipedia (see: http://en.wikipedia.org/wiki/Representational_state_transfer), here are the following definitions for POST and PUT to an Element URI (e.g. /drafts/<draft_id>):

PUT:
Replace the addressed member of the collection, or if it doesn't exist, create it.

POST:
Not generally used. Treat the addressed member as a collection in its own right and create a new entry in it.

Technically, we are abusing 'PUT' as an 'UPDATE' call, which is certainly bending the RESTful semantics (see: http://jcalcote.wordpress.com/2008/10/16/put-or-post-the-rest-of-the-story/ , http://cafe.elharo.com/web/put-is-not-update/ ). However, we are using the PUT method in other places within the project to mean UPDATE (it's even called draft_update_api). Therefore, we should at least be consistent and use the PUT method on drafts for updating them.

Fixes T360

Test Plan: Review code, references, run tests.

Reviewers: spang

Reviewed By: spang

Maniphest Tasks: T360

Differential Revision: https://review.inboxapp.com/D391