[namespaces] Add filtering to the namespaces.
This makes sense as we have filtering on all other endpoints of this form. It
is also needed by the cliend sdks which currently always provide an offset and
limit parameter when grabbing collections.

Summary:
...ns. It is also needed for the test
client which will want to filter by an email address to interact with a
particular account when one of many has been authenticated (yet there is no
filtering by say access_token in non-hosted version.)

This makes sense as we have filtering on all other endpoints of this form. It
is also needed by the cliend sdks which currently always provide an offset and
limit parameter when grabbing collections.

It is also needed for the test
client which will want to filter by an email address to interact with a
particular account when one of many has been authenticated (yet there is no
filtering by say access_token in non-hosted version.)

Add Filtering to the namespaces.

Fixes T382

Test Plan: Run new tests that exercise the various filtering parameters.

Reviewers: kav-ya

Maniphest Tasks: T382

Differential Revision: https://review.inboxapp.com/D405