[tests/system] Updating the system test client to use the python-sdk
Summary:
This basically just converts the old system tests to use the new python-sdk.
There's enough code different that it may not make sense to look too much at
the changes but instead focus on what the new code looks like. Since this code
doesn't actually touch any of the main system, I didn't run the unit tests.
Furthermore, I'm not sure the old stuff works but the new stuff certainly works
*better*. Note that not all of the tests pass, but I think it's better to land
this framework and then focus on those tests ass different diffs.

In particular, I think you'll probably benefit from looking at the actual system
tests to see how they're laid out now so that we can start creating tests that
use this system to test more complex behavior.

Note that to run the tests, you will need to have an 'accounts.py' file in the
`tests/system` directory with a format as defined in
`tests/system/conftest.py`. When we run into issues in the future, I think we
should push hard to create system tests that reproduce the issue first so that
we can verify it as a) reproducible and b) fixed when the test passes and c)
ensure that the issue doesn't arise again in the future.

One thing that is missing right now is some framework for actually starting up
and stopping all of the services. Again, I am going to work on that (probably
next) in a separate diff so that running all of the system tests can be
automated (as you know, you manually need to start all of the services right
now... and as discussed at lunch, doing anything twice that can be automated is
a waste of time).

Maybe the title of this section should be called "Description" rather than
"Summary" as this wall of text was perhaps a bit long-winded. If you've made it
this far. Cheers!

* [tests/sending] Use decorators to handle timeout loops.
* [tests/system/send] Fix a small typo.
* make test_sending pass
* working drafts test
* port drafts to new client
* removed redundant test
* [tests/system] More system test improvements.

Test Plan: review the new code.

Reviewers: kav-ya

Differential Revision: https://review.inboxapp.com/D415